+++ src/module/clearAll.js
@@ -1,4 +1,4 @@
-import { indexUpdate } from './updateIndex.js';
+import indexUpdate from './updateIndex.js';
 import display from './display.js';
 import getFromStorage, { saveInStorage } from './storage.js';
